Shared function to process link wallet logic
